Feat: 유저 정보 입력 구현

+++ src/main/java/cat/soft/src/parking/model/ParkingLot.java
@@ -13,5 +13,5 @@ public class ParkingLot {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "idx", nullable = false)
-	private long idx;
+	private Integer idx;
 }
